% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCRFileTransformation.R
\name{addThresholdCq}
\alias{addThresholdCq}
\title{Add systemCalculatedThresholdValue, systemCalculatedCqValue, and userProvidedCqValue to
results_Table or standardCurveResults_Table files.}
\usage{
addThresholdCq(file, calculateUserProvidedCq)
}
\arguments{
\item{file}{results_Table or standardCurveResults_Table file that is read in.}

\item{calculateUserProvidedCq}{defines if user wants UserProvidedCq value to be calculated and
added to file. Update variable to "Yes" to have value added or "No" if not.}
}
\value{
Writes updated CSV file to local directory.
}
\description{
Function adds systemCalculatedThresholdValue and systemCalculatedCqValue to
results_Table or standardCurveResults_Table files. User can also add userProvidedCqValue to
file by setting calculateUserProvidedCq parameter to "Yes". If user wants system to calculate
userProvidedCqValue they must provide userProvidedThresholdValue's.
}
