% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCRFileTransformation.R
\name{formatRawFluorescenceFile}
\alias{formatRawFluorescenceFile}
\title{Format raw qPCR fluorescence file data for results_Table or standardCurveResults_Table.}
\usage{
formatRawFluorescenceFile(rawFluorescenceFile, platform, outputFileName)
}
\arguments{
\item{rawFluorescenceFile}{defines raw qPCR fluorescence file to read in.}

\item{platform}{defines platform fluorescence file was generated from.}

\item{outputFileName}{defines output name for CSV file that data will be written to.}
}
\value{
Writes CSV file to local directory.
}
\description{
Function formats raw qPCR fluoresence file data from MIC, StepOnePlus or
Biomeme two3/Franklin machines into table that includes well location names and fluorescence
data for each reaction cycle. The table is written to the local machine directory as a CSV
file. The formatted data can be copied and pasted into the results_Table or
standardCurveResults_Table, which are used in the MDMAPR 2.0 MySQL database.
}
