% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch the MDMAPR app}
\usage{
launchApp()
}
\value{
shiny application object
}
\description{
This function runs the MDMAPR Shiny web application. Please run the
dbInstance  function first to specify if you are running the application with or with
out a database. If you are running the application with a databse, then after the dbInstance function run the dbVariables function to input the variabels needed to
establish a  connection to your MDMAPR MySQL database instance. Then run launchApp().
}
