% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAppUI.R
\docType{data}
\name{shinyAppUI}
\alias{shinyAppUI}
\title{MDMAPR UI Function}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
shinyAppUI
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}
}
\description{
MDMAPR UI Function
}
\keyword{datasets}
