% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAppUI.R
\docType{data}
\name{UI}
\alias{UI}
\alias{css}
\title{MDMAPR UI Function}
\format{
An object of class \code{html} (inherits from \code{character}) of length 1.
}
\usage{
css
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}
}
\description{
MDMAPR UI Function
}
\keyword{datasets}
