% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInstance.R
\name{dbInstance}
\alias{dbInstance}
\title{Indicate whether MDMAPR is connected to database instance.}
\usage{
dbInstance(x)
}
\arguments{
\item{x}{defines if there is a database connection or not}
}
\value{
"Yes" or "No"
}
\description{
Function to indicate if MDMAPR will be run with a database instance.
If the application is run with a database, update dbVariables function.
}
\details{
dbInstance
}
\examples{
dbInstance("Yes")
dbInstance("No")

}
