% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInstance.R
\name{dbInstance}
\alias{dbInstance}
\title{Indicate whether MDMAPR is connected to database instance.}
\usage{
dbInstance(x)
}
\arguments{
\item{x}{defines if you have a database connection or not}
}
\value{
"Yes" or "No"
}
\description{
Function to indicate if you are running MDMAPR with a database instance.
}
\details{
dbInstance
}
\examples{
"Yes" or "No"

}
