% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbVariables.R
\name{dbVariables}
\alias{dbVariables}
\title{Information needed to connect to local or remote database instance.}
\usage{
dbVariables(user = "root", password = "Test123!", dbname = 'MDMap_2.0', host = "127.0.0.1")
}
\arguments{
\item{user}{defines username for database connection.}

\item{password}{define password for database connection.}

\item{dbname}{defines name of database instance.}

\item{host}{defines host name for database connection.}
}
\description{
function stores username, password, database name, and host name for database instance to connect to.
}
\details{
dbVariables
}
